[sr] Respect the uniqueness of SPIR-V types
The spec says that "Non-structure types (scalars, vectors, arrays,
etc.) with the same operand parameterization cannot be type aliases."
So we need to check whether the type to be created already exits.
Right now we do a linear search in the vector.

Also create the token class for SPIR-V constants.