Parse header version from the binary
Currently the version doesn't appear to be read from the header
in parse_header. So I've updated parse_header to parse the version
and reused the same Word -> (u8, u8) function which already
existed (just moved it to a separate module).

This fixes the example in the README (currently assert_eq fails)
and some tests/doctests which seemed incorrect previously.