sr: Centralized operand tokenization
Instead of having different methods for composing specific tokens for
an operand in SR autogenerator, this PR introduces a structure that contains
all the tokens associated/needed for an operand.

This helps to keep the logic consistent across different usages of an
operand: e.g. between the code that defines it as a field versus the code
that parses a dr::Instruction to produce the value.