autogen: Convert from build script to runnable binary
According to the autogen README rspirv-autogen is used to "automatically
update files in anther crates". This use seems long-gone as none of the
top-level projects depend on `rspirv-autogen`: regeneration does not
happen at build-time anymore.

However, when loading this project in an IDE that scans and `check`s all
workspaces the build script _is_ invoked resulting in slow startup
times and accidental regeneration, which is particularly inconvenient
when SPIRV-Headers so happens to be checked out on a different commit.