binary: Only parse parameters for overlapping operands once
If an operand has two distinct variants with the same value a SPIR-V
word would be consumed and decoded into an operand more than once,
breaking the remaining parsing code.  This happens for example with
MAKE_POINTER_AVAILABLE and MAKE_POINTER_AVAILABLE_KHR which have the
same value and would call the parameter parsing code twice.