spirv: Drop stale "tracking SPIR-V major/minor version" from README
Since inheriting the `spirv` crate and dropping the `spirv_headers`
crate in #204, and following up on a choice in #197 to no longer have
the SPIR-V major/minor version in our crate version which disallows us
from making any breaking changes to the crate, we reset the version to
`0.1.0` and embedded the SPIR-V version via _version metadata_ instead.
This stale comment in the README was still indicating as such though,
confusing users in e.g. #252 that our `spirv` crate was somehow exposing
SPIR-V 1.3 (should have been 0.3 by that logic which is the current
latest version).  Remove it entirely.

Note also that since #225 / #226 our version metadata is no longer the
SPIR-V version/revision but the Vulkan SDK tag that it was released
with.  The SPIR-V version isn't bumped often enough to match extensions
in new SDK releases, making the SDK tag more indicative of the included
API surface instead.